Merges gadget.cpp/h into common from redalert.
