+++ tiberiandawn/CMakeLists.txt
@@ -45,7 +45,6 @@ set(TIBDAWN_SRC
     fly.cpp
     foot.cpp
     fuse.cpp
-    gadget.cpp
     gamedlg.cpp
     gauge.cpp
     globals.cpp
